Remove useless _allgather_then_aggregate_hook.
This hook is not used at all, and it probably can only be used for demonstrating that allgather is slower than allreduce, so it should never be used in practice.

However, this hook and its helper function stay with the communication hook public APIs in the same file. It will be better to make the public API file as concise as possible.

Since I don't think we will use this hook in the future, prefer deleting it to moving it to a separate file.

Differential Revision: [D26575318](https://our.internmc.facebook.com/intern/diff/D26575318/)

ghstack-source-id: 122180969
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52593